+++ include/particleintegrals/contract/direct4C_libcint_spindependent.hpp
@@ -643,7 +643,7 @@ namespace ChronusQ {
 #endif
 
 #ifdef _REPORT_INTEGRAL_TIMINGS
-      size_t nIntSkip = std::accumulate(nSkip.begin(),nSkip.end(),0);
+      size_t nIntSkip = std::accumulate(nSkip.begin(),nSkip.end(),size_t(0));
       std::cout << "Bare-Coulomb-Exchange Libcint Screened " << nIntSkip << std::endl;
   
       auto durDirect = tock(topDirect);
@@ -1119,7 +1119,7 @@ namespace ChronusQ {
 #endif
 
 #ifdef _REPORT_INTEGRAL_TIMINGS
-      size_t nIntSkipLL = std::accumulate(nSkipLL.begin(),nSkipLL.end(),0);
+      size_t nIntSkipLL = std::accumulate(nSkipLL.begin(),nSkipLL.end(),size_t(0));
       std::cout << "Dirac-Coulomb-C(2) Screened " << nIntSkipLL << std::endl;
 
       auto durDirectLL = tock(topDirectLL);
@@ -1996,7 +1996,7 @@ namespace ChronusQ {
 #endif
 
 #ifdef _REPORT_INTEGRAL_TIMINGS
-      size_t nIntSkipSSSS = std::accumulate(nSkipSSSS.begin(),nSkipSSSS.end(),0);
+      size_t nIntSkipSSSS = std::accumulate(nSkipSSSS.begin(),nSkipSSSS.end(),size_t(0));
       std::cout << "Dirac-Coulomb-SSSS Screened " << nIntSkipSSSS << std::endl;
   
       auto durDirectSSSS = tock(topDirectSSSS);
@@ -2936,7 +2936,7 @@ namespace ChronusQ {
 #endif
   
 #ifdef _REPORT_INTEGRAL_TIMINGS
-      size_t nIntSkipGaunt = std::accumulate(nSkipGaunt.begin(),nSkipGaunt.end(),0);
+      size_t nIntSkipGaunt = std::accumulate(nSkipGaunt.begin(),nSkipGaunt.end(),size_t(0));
       std::cout << "Gaunt Screened " << nIntSkipGaunt << std::endl;
   
       auto durDirectGaunt = tock(topDirectGaunt);
@@ -3918,7 +3918,7 @@ namespace ChronusQ {
 #endif
   
 #ifdef _REPORT_INTEGRAL_TIMINGS
-      size_t nIntSkipGauge = std::accumulate(nSkipGauge.begin(),nSkipGauge.end(),0);
+      size_t nIntSkipGauge = std::accumulate(nSkipGauge.begin(),nSkipGauge.end(),size_t(0));
       std::cout << "Gauge Screened " << nIntSkipGauge << std::endl;
   
       auto durDirectGauge = tock(topDirectGauge);
